backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
